+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## 0.8.1
+
+  - Breaking change: Content=>List<Part>, lots more List<> => Iterable<>
+
 ## 0.8.0
 
 - Multimedia Input Support: Added `attachments` parameter to Agent and Model
   interfaces for including files, data and links.
-- Improvided OpenAI compatibiliy for tool calls
+- Improved OpenAI compatibility for tool calls
 - Added the 'gemini-compat' provider for access to Gemini models via the
   OpenAI endpoint.
 - Breaking change: everywhere I passed List<Message> I now pass